requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
openpyxl>=3.1.2
pydantic>=2.7.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
httpx>=0.27.0

# OCR Ön-filtre (2-Aşamalı Doğrulama)
Pillow>=10.0.0
pytesseract>=0.3.10

